todolist/urls.py
from django.contrib import admin
from django.urls import path, include
from tasks.api import TaskViewSet
from profiles.views import ProfileCreationView
from rest_framework import routers
from rest_framework.authtoken import views
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)

urlpatterns = [
    path('admin/', admin.site.urls),
    path('', include('tasks.urls')),

    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path("api/profiles/", ProfileCreationView.as_view(), name="profiles"),
]
